separate docking base schema from posit specifics
